pyproject.toml
[tool.poetry]
name = "kubernetes-debugger"
version = "0.1.0"
description = "LLM-powered Kubernetes cluster debugger"
authors = ["Chris Sanders <sanders.chris@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "*"
anthropic = "^0.37.1"
openai = "^1.52.1"
#openai = {extras = [], version = "^1.52.1"}

[tool.poetry.group.dev.dependencies]
#black = "^23.10.0"
#pylint = "^3.0.2"
#pytest = "^7.4.3"

[build-system]
#requires = ["poetry-core>=1.0.0"]
#build-backend = "poetry.core.masonry.api"
